# openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/usuarios/editar/{id}:
    put:
      tags:
      - usuario-controller
      operationId: actualizarUsuario
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UsuarioDTO"
  /api/usuarios/guardar:
    post:
      tags:
      - usuario-controller
      operationId: guardarUsuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UsuarioDTO"
  /api/usuarios:
    get:
      tags:
      - usuario-controller
      operationId: obtenerTodosUsuarios
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UsuarioDTO"
  /api/usuarios/eliminar/{id}:
    delete:
      tags:
      - usuario-controller
      operationId: eliminarUsuario
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UsuarioDTO"
components:
  schemas:
    UsuarioDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombres:
          type: string
        apellidos:
          type: string
        rut:
          type: integer
          format: int64
        dv:
          type: string
        fechaNacimiento:
          type: string
          format: date-time
        correoElectronico:
          type: string
        contrasena:
          type: string